requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
google-cloud-bigquery==3.13.0
pandas==2.1.3
python-multipart==0.0.6
